0413
서버 수업 내용 정리 및 복습

TCP를 이용하여 액터 위치 및 회전 동기화

PlayerCharacter 이동 시 UpdateTick에 따라 UpdateMessage BroadCast

서버는 받은 패킷 정보를 다른 유저들에게 보내주며 해당 클라이언트들은 받은 정보를 이용해 OtherPlayerCharacter의 위치 업데이트

+++ Source/HonorProject/GameMode/PlayGameMode.h
@@ -30,7 +30,7 @@ private:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "ClientData", meta = (AllowPrivateAccess = "true"))
 	TArray<TSubclassOf<AMasterAICharacter>> m_AICharacterClasses;
 
-	TMap<int, ServerObject>					m_AllObject;
+	TMap<uint64_t, ServerObject>			m_AllObject;
 	TMap<EGameObjectType, TArray<AActor*>>	m_TypeOfAllObject;
 
 	int m_ClientUniqueID;
